Guards

+++ server/controllers/SubjectController.go
@@ -2,8 +2,11 @@ package controllers
 
 import (
 	"edumeet/dtos"
+	"edumeet/ent"
+	"edumeet/guards"
 	"edumeet/services"
 
+	"github.com/go-playground/validator/v10"
 	"github.com/gofiber/fiber/v2"
 	"github.com/oklog/ulid/v2"
 )
@@ -23,9 +26,23 @@ func NewSubjectController(subjectService *services.SubjectService, emailService 
 func (sc *SubjectController) Create(c *fiber.Ctx) error {
 
 	var subjectDTO dtos.SubjectDTO
+	currentUser := c.Locals("user").(*ent.User)
+	if !guards.IsAdmin(currentUser) {
+		return c.Status(fiber.StatusForbidden).JSON(fiber.Map{"error": "Not authorized"})
+	}
 
 	if err := c.BodyParser(&subjectDTO); err != nil {
 		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": "Invalid request"})
+	}
+
+	validations := validator.New()
+	err := validations.Struct(subjectDTO)
+	if err != nil {
+		errors := make([]string, 0)
+		for _, err := range err.(validator.ValidationErrors) {
+			errors = append(errors, err.Error())
+		}
+		return c.Status(fiber.StatusUnprocessableEntity).JSON(fiber.Map{"error": errors})
 	}
 
 	subject, err := sc.subjectService.Create(subjectDTO)
@@ -50,6 +67,16 @@ func (sc *SubjectController) Update(c *fiber.Ctx) error {
 		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": "Invalid request"})
 	}
 
+	validations := validator.New()
+	err = validations.Struct(subjectDTO)
+	if err != nil {
+		errors := make([]string, 0)
+		for _, err := range err.(validator.ValidationErrors) {
+			errors = append(errors, err.Error())
+		}
+		return c.Status(fiber.StatusUnprocessableEntity).JSON(fiber.Map{"error": errors})
+	}
+
 	subject, err := sc.subjectService.Update(id.String(), subjectDTO)
 
 	if err != nil {
@@ -65,6 +92,11 @@ func (sc *SubjectController) Delete(c *fiber.Ctx) error {
 	if err != nil {
 		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": "Invalid ID"})
 	}
+	currentUser := c.Locals("user").(*ent.User)
+	if !guards.IsAdmin(currentUser) {
+		return c.Status(fiber.StatusForbidden).JSON(fiber.Map{"error": "Not authorized"})
+	}
+
 	errDelete := sc.subjectService.Delete(id.String())
 
 	if errDelete != nil {
